core/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "infimum-core"
version = "v1.2.11"
description = "Infimum core library: base, engine, database, AI (LLM/VLM/speech), utils"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "decorator>=5.2.0",
    "loguru>=0.7.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.0",
    "pytz",
    "redis>=7.1.0",
    "requests",
    "beautifulsoup4",
    "sqlalchemy>=2.0.45",
    "openai>=2.15.0",
    "numpy>=1.24.0",
    "pillow>=11.0",
    "opencv-python>=4.8.0",
    "librosa>=0.11.0",
]

[project.optional-dependencies]
# FastAPI support (for api_router_registry and web apps)
api = [
    "fastapi>=0.128.0",
]
# MongoDB support (sync and async)
mongo = [
    "pymongo>=4.16.0",
    "motor>=3.6.0",
]

# Milvus vector database support
milvus = [
    "pymilvus>=2.6.6",
]

# Qdrant vector database support
qdrant = [
    "qdrant-client>=1.7.0",
]

# All database backends
all-db = [
    "pymongo>=4.16.0",
    "motor>=3.6.0",
    "pymilvus>=2.6.6",
    "qdrant-client>=1.7.0",
]

# Security module support (JWT, password hashing)
security = [
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
]

# LLM/agent support (LangChain, LangGraph)
llm = [
    "langchain-core>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.2.0",
]

# Build from core/; discover package "core" and subpackages in parent (repo root)
[tool.setuptools.packages.find]
where = [".."]
include = ["core*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
